Moving indices(of:) to RandomAccessCollectionExtensions
