+++ Sources/Extensions/SwiftStdlib/ArrayExtensions.swift
@@ -179,22 +179,6 @@ public extension Array {
 // MARK: - Methods (Equatable)
 public extension Array where Element: Equatable {
 
-	/// SwifterSwift: All indices of specified item.
-	///
-	///		[1, 2, 2, 3, 4, 2, 5].indices(of 2) -> [1, 2, 5]
-	///		[1.2, 2.3, 4.5, 3.4, 4.5].indices(of 2.3) -> [1]
-	///		["h", "e", "l", "l", "o"].indices(of "l") -> [2, 3]
-	///
-	/// - Parameter item: item to check.
-	/// - Returns: an array with all indices of the given item.
-	public func indices(of item: Element) -> [Index] {
-		var indices: [Index] = []
-		for index in startIndex..<endIndex where self[index] == item {
-			indices.append(index)
-		}
-		return indices
-	}
-
 	/// SwifterSwift: Remove all instances of an item from array.
 	///
 	///		[1, 2, 2, 3, 4, 5].removeAll(2) -> [1, 3, 4, 5]
